Updating API calls.

+++ server/src/core/methods/wait.c
@@ -28,6 +28,7 @@
  */
 
 #include "private.h"
+#include <lib3270/keyboard.h>
 
 int ipc3270_method_wait(GObject *session, GVariant *request, GObject *response, GError G_GNUC_UNUSED(**error)) {
 
@@ -60,7 +61,7 @@ int ipc3270_method_wait_for_unlock(GObject *session, GVariant *request, GObject 
 
 	guint seconds = 1;
 	g_variant_get(request, "(u)", &seconds);
-	ipc3270_response_append_int32(response, lib3270_wait_for_unlock(ipc3270_get_session(session),seconds));
+	ipc3270_response_append_int32(response, (int32_t) lib3270_wait_for_unlock(ipc3270_get_session(session),seconds));
 	return 0;
 
 }
